requirements.txt
streamlit
requests
python-dotenv
google-generativeai
pydantic
